\name{harbin-package}
\alias{harbin-package}
\alias{harbin}
\docType{package}
\title{
Harbin quantification for RT-qPCR data
}
\description{Performs Harbin quantification for reverse transcription quantitative PCR (RT-qPCR) data.
}
\details{
\tabular{ll}{
Package: \tab harbin\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2015-02-26\cr
License: \tab GPL-2\cr
}
}
\author{
Theo Pepler

Maintainer: Theo Pepler <tpepler@sun.ac.za>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link{harbin.quant}}
}
\examples{
# To start the Harbin quantification programme, type

harbin.quant()

# at the R prompt, hit Enter, and follow the instructions.
}
