\name{harbin-package}
\alias{harbin-package}
\alias{harbin}
\docType{package}
\title{
Harbin quantification for RT-qPCR data
}
\description{Performs Harbin quantification for reverse transcription quantitative PCR (RT-qPCR) data. Currently offers support only for files produced by the Qiagen Rotor-Gene Q thermal cycler, but the package can be extended per request to support other technologies.
}
\details{
\tabular{ll}{
Package: \tab harbin\cr
Type: \tab Package\cr
Version: \tab 0.1-4\cr
Date: \tab 2015-09-07\cr
License: \tab GPL-2\cr
}
}
\author{
Theo Pepler

Maintainer: Theo Pepler <theo.pepler@glasgow.ac.uk>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link{harbin.quant}}
}
\examples{
# To start the Harbin quantification programme, type

harbin.quant()

# at the R prompt, hit Enter, and follow the instructions.
}
